requirements.txt
# Python dependencies for TradingView → Telegram Alert Bot
# Install with: pip install -r requirements.txt

fastapi==0.111.0          # Web framework for the webhook server
uvicorn[standard]==0.30.1 # ASGI server to run FastAPI
httpx==0.27.0             # Async HTTP client for Telegram API calls
python-dotenv==1.0.1      # Loads .env file into environment variables
